rewrite based on input from u/Mlakuss on reddit
There are two very important concepts I was previously unaware of:

 1. Any entity can have a scoreboard, not just players.
 2. You can "tag" entities for reference later.

This means, instead of grouping all non-powers-of-two into
powers-of-two, we can instead select a random orb, and add up all nearby
orbs.

This is complicated by the fact that, after removing a "merged" XP orb
(ie: the kill command), already-queued iterations from

    execute as @e[type=minecraft:experience_orb] ...

will still be processed. To avoid this, we add a "dead" tag to anything
we are about to kill, then check for that before entering the body of
the iteration.